+++ test/oolong/pick_deep_test.js
@@ -8,14 +8,14 @@ describe("Oolong.pickDeep", function() {
 
   it("must return a new object with given 2nd level key", function() {
     var person = {name: "Alice", address: {country: "UK", street: "Downing"}}
-    var obj = O.pickDeep(person, "name", "address.country")
+    var obj = O.pickDeep(person, "name", ["address", "country"])
     obj.must.eql({name: "Alice", address: {country: "UK"}})
   })
 
   it("must return a new object with given 3rd level key", function() {
     var children = {Bob: {name: "Bob", age: "13"}}
     var person = {name: "Alice", children: children}
-    var obj = O.pickDeep(person, "name", "children.Bob.name")
+    var obj = O.pickDeep(person, "name", ["children", "Bob", "name"])
     obj.must.eql({name: "Alice", children: {Bob: {name: "Bob"}}})
   })
 
@@ -27,7 +27,10 @@ describe("Oolong.pickDeep", function() {
   it("must return a new object with given 2nd level keys", function() {
     var address = {country: "UK", street: "Downing", apartment: 10}
     var person = {name: "Alice", address: address}
-    var obj = O.pickDeep(person, "name", "address.country", "address.street")
+    var obj = O.pickDeep(person,
+                         "name",
+                         ["address", "country"],
+                         ["address", "street"])
     obj.must.eql({name: "Alice", address: {country: "UK", street: "Downing"}})
   })
 
@@ -54,14 +57,14 @@ describe("Oolong.pickDeep", function() {
 
   it("must not return non-existent 2nd level keys", function() {
     var person = {name: "Alice", address: {country: "UK", street: "Downing"}}
-    var obj = O.pickDeep(person, "name", "address.apartment")
+    var obj = O.pickDeep(person, "name", ["address", "apartment"])
     obj.must.eql({name: "Alice"})
   })
 
   it("must not return non-existent 3rd level keys", function() {
     var children = {Bob: {name: "Bob", age: "13"}}
     var person = {name: "Alice", children: children}
-    var obj = O.pickDeep(person, "name", "children.Bob.height")
+    var obj = O.pickDeep(person, "name", ["children", "Bob", "height"])
     obj.must.eql({name: "Alice"})
   })
 })